gdg-certificate-generator/requirements.txt
streamlit
Pillow
pandas
openpyxl
qrcode[pil]
img2pdf
requests
